back-end/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.0
fonttools==4.39.3
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
passlib==1.7.4
Pillow==9.5.0
pipreqs==0.4.13
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
UliPlot==0.2.2
urllib3==1.26.15
uvicorn==0.21.1
XlsxWriter==3.1.0
xlwt==1.3.0
yarg==0.1.9
zipp==3.15.0
